refactor PlayMapScene and WorldMapScene to MapScene
