+++ mygame/app/requires.rb
@@ -14,6 +14,7 @@ require "#{File.dirname(__FILE__)}/game/gui/gui_elements/button.rb"
 require "#{File.dirname(__FILE__)}/game/gui/button_defs.rb"
 require "#{File.dirname(__FILE__)}/game/scenes/scene_base.rb"
 require "#{File.dirname(__FILE__)}/game/scenes/main_menu_scene.rb"
+require "#{File.dirname(__FILE__)}/game/scenes/map_scene.rb"
 require "#{File.dirname(__FILE__)}/game/scenes/world_map_scene.rb"
 require "#{File.dirname(__FILE__)}/game/scenes/play_map_scene.rb"
 require "#{File.dirname(__FILE__)}/game/models/noise/perlin_noise.rb"
